+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/main.js
@@ -61,6 +61,9 @@ define([
             IbanExtension.initialize(app);
             ContactButtons.initialize(app);
 
+            sandbox.urlManager.setUrl('contact', 'contacts/contacts/edit:<%= id %>/details');
+            sandbox.urlManager.setUrl('account', 'contacts/accounts/edit:<%= id %>/details');
+
             Config.set('sulucontact.components.autocomplete.default.contact', {
                 remoteUrl: '/admin/api/contacts?searchFields=id,fullName&flat=true&fields=id,fullName&limit=25',
                 getParameter: 'search',